jenkins CI script

+++ webscrab.py
@@ -19,23 +19,34 @@ class MyParser(HTMLParser):
         if tag == 'a':
             for name, value in attrs:
                 if name == 'href':
-                    self.resault = value  # merge with latest apk
-                    # print  value
+                    self.resault = value  # merge with the latest apk
 
     def getHtml(self, url):
         page = urllib.urlopen(url)
         html = page.read()
-        # print  html
         return html
 
     def startAppium(self):
         # cmd = 'start /b appium'
         # cmd = 'nohup appium & '
+        # cmd = 'echo `nohup /usr/local/bin/appium &` &'
         cmd = 'date'
         print "-------%s-------" % ctime()
         print os.popen(cmd).read()
 
-    def job_1(self):
+    def job_Mac(self):
+        # cmd = "rm ./*"
+        # print os.popen(cmd).read()
+
+        cmd = " git clone https://git.hz.netease.com/git/hzxiadaqiang/Script.git"
+        print os.popen(cmd).read()
+
+        cmd ="cp -r ./Script/*   ./  && rm -r ./Script "
+        print os.popen(cmd).read()
+
+        cmd = "mkdir RFUI_outputs_dir"
+        print os.popen(cmd).read()
+        cmd = "touch ./RFUI_outputs_dir/log.html ./RFUI_outputs_dir/"
 
         cmd = "git clone https://git.hz.netease.com/git/yxplusQA/YX_RFUI_Framework_demo.git"
         print os.popen(cmd).read()
@@ -50,15 +61,65 @@ class MyParser(HTMLParser):
         print self.resault
         apk_version = self.resault
         apk_version = apk_version.replace('/', '')
-        download_url = "http://10.240.129.99/nightly/" + apk_version + '/' + apk_version + ".apk"
-        f = urllib2.urlopen(download_url)
-        self.pwd = self.get_pwd()
-        print self.pwd
-        data = f.read()
-        filename = apk_version + ".apk"
-        with open(filename, "wb") as code:
-            code.write(data)
-        print "Download Done!"
+        print apk_version
+
+        if  os.path.exists('./'+apk_version+'.apk'):
+            print "the Apk already up-to-date\n"
+            sleep(8) # wait appium start
+        else:
+            download_url = "http://10.240.129.99/nightly/" + apk_version + '/' + apk_version + ".apk"
+            f = urllib2.urlopen(download_url)
+            self.pwd = self.get_pwd()
+            print self.pwd
+            data = f.read()
+            filename = apk_version + ".apk"
+            with open(filename, "wb") as code:
+                code.write(data)
+            print "Download Done!"
+
+    def job_Windows(self):
+        # cmd = "rm ./*"
+        # print os.popen(cmd).read()
+
+        cmd = " git clone https://git.hz.netease.com/git/hzxiadaqiang/Script.git"
+        print os.popen(cmd).read()
+
+        cmd ="cp -r ./Script/*   ./  && rm -r ./Script "
+        print os.popen(cmd).read()
+
+        cmd = "mkdir RFUI_outputs_dir"
+        print os.popen(cmd).read()
+        cmd = "touch ./RFUI_outputs_dir/log.html ./RFUI_outputs_dir/"
+
+        cmd = "git clone https://git.hz.netease.com/git/yxplusQA/YX_RFUI_Framework_demo.git"
+        print os.popen(cmd).read()
+
+        # cmd = "wget https://git.hz.netease.com/hzxiadaqiang/Script/blob/master/webscrab.py?raw=true && mv webscrab.py?raw=true webscrab.py"
+        # print os.popen(cmd).read()
+
+        html = self.getHtml('http://10.240.129.99/nightly/')
+        apk_list = self.feed(html)
+        print "-------%s-------" % ctime()
+        print apk_list
+        print self.resault
+        apk_version = self.resault
+        apk_version = apk_version.replace('/', '')
+        print apk_version
+
+        if  os.path.exists('./'+apk_version+'.apk'):
+            print "the Apk already up-to-date\n"
+            sleep(8) # wait appium start
+        else:
+            download_url = "http://10.240.129.99/nightly/" + apk_version + '/' + apk_version + ".apk"
+            f = urllib2.urlopen(download_url)
+            self.pwd = self.get_pwd()
+            print self.pwd
+            data = f.read()
+            filename = apk_version + ".apk"
+            with open(filename, "wb") as code:
+                code.write(data)
+            print "Download Done!"
+
         # rm the old apk
         cmd = "rm ./YX_RFUI_Framework_demo/Resources/yixin_test.apk"
         print os.popen(cmd).read()
@@ -66,48 +127,19 @@ class MyParser(HTMLParser):
         cmd = "ln -s ./" + apk_version + ".apk" + "   ./YX_RFUI_Framework_demo/Resources/yixin_test.apk"
         print os.popen(cmd).read()
 
-        cmd = "pybot --variable BROWSER:safari --outputdir safari_dir --include demo --xunit output_xunit.xml --xunitskipnoncritical ./YX_RFUI_Framework_demo/Test/YX_Subscriptions/test_suite_examples.txt"
+        cmd = "pybot --variable BROWSER:safari --outputdir safari_dir --include Androiddemo --xunit output_xunit.xml --xunitskipnoncritical ./YX_RFUI_Framework_demo/Test/YX_Subscriptions/test_suite_examples.txt"
         # cmd = "ps aux"
         # print "[%s]:" %ctime()
         print "-------%s-------" % ctime()
         print os.popen(cmd).read()
 
 
-# if __name__ ==  ' __main__ ' :    <------ 该冒号为中文 导致run不会执行
 if __name__ == '__main__':
 
     MyParser = MyParser()
-    """
-    html = MyParser.getHtml('http://10.240.129.99/nightly/')
-    # print html
-    apk_list = MyParser.feed(html)
-    # print type(apk_list)
-    print apk_list
-    print MyParser.resault
-    apk_version = MyParser.resault
-    apk_version = apk_version.replace('/', '')
-    # MyParser.feed(html)
-    download_url = "http://10.240.129.99/nightly/" + apk_version + '/' + apk_version + ".apk"
-    # print  download_url
-    f = urllib2.urlopen(download_url)
-    MyParser.pwd = MyParser.get_pwd()
-    print MyParser.pwd
-
-    # cmd = "pybot --variable BROWSER:safari --outputdir safari_dir --include demo --xunit output_xunit.xml --xunitskipnoncritical Test/YX_Subscriptions/test_suite_examples.txt"
-    # print os.popen(cmd).read()
-
-
-    data = f.read()
-    filename= apk_version+".apk"
-    with open(filename, "wb") as code:
-        code.write(data)
-    print "Download Done!"
-
-    MyParser.startAppium()
-    """
     threads = []
     t1 = threading.Thread(target=MyParser.startAppium())
-    t2 = threading.Thread(target=MyParser.job_1())
+    t2 = threading.Thread(target=MyParser.job_Mac())
     threads.append(t1)
     threads.append(t2)
     for t in threads:
